# aidguide_04_ws/src/aidguide_04_backend/api/swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "AidGuide API",
      "description": "API para la aplicación AidGuide que permite la gestión de usuarios, robots, rutas y más",
      "version": "1.0.0",
      "contact": {
        "email": "contacto@aidguide.com"
      }
    },
    "basePath": "/api",
    "schemes": ["http", "https"],
    "consumes": ["application/json"],
    "produces": ["application/json"],
...
    "paths": {
...
      "/usuarios": {
        "get": {
          "tags": ["usuarios"],
          "summary": "Obtiene todos los usuarios",
          "operationId": "getUsuarios",
          "produces": ["application/json"],
          "responses": {
            "200": {
              "description": "Lista de usuarios",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Usuario"
                }
              }
            },
            "500": {
              "description": "Error del servidor",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        },
        "post": {
          "tags": ["usuarios"],
          "summary": "Crea un nuevo usuario",
          "operationId": "createUsuario",
          "produces": ["application/json"],
          "consumes": ["application/json"],
          "parameters": [
            {
              "in": "body",
              "name": "usuario",
              "description": "Datos del usuario a crear",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Usuario"
              }
            }
          ],
          "responses": {
            "201": {
              "description": "Usuario creado",
              "schema": {
                "$ref": "#/definitions/Usuario"
              }
            },
            "400": {
              "description": "Datos de entrada inválidos",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            },
            "500": {
              "description": "Error del servidor",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },
      "/usuarios/{id}": {
        "get": {
          "tags": ["usuarios"],
          "summary": "Obtiene un usuario por ID",
          "operationId": "getUsuarioById",
          "produces": ["application/json"],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID del usuario",
              "required": true,
              "type": "integer",
              "format": "int64"
            }
          ],
          "responses": {
            "200": {
              "description": "Usuario encontrado",
              "schema": {
                "$ref": "#/definitions/Usuario"
              }
            },
            "404": {
              "description": "Usuario no encontrado",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            },
            "500": {
              "description": "Error del servidor",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        },
        "put": {
          "tags": ["usuarios"],
          "summary": "Actualiza un usuario existente",
          "operationId": "updateUsuario",
          "produces": ["application/json"],
          "consumes": ["application/json"],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID del usuario a actualizar",
              "required": true,
              "type": "integer",
              "format": "int64"
            },
            {
              "in": "body",
              "name": "usuario",
              "description": "Datos actualizados del usuario",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Usuario"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Usuario actualizado",
              "schema": {
                "$ref": "#/definitions/Usuario"
              }
            },
            "404": {
              "description": "Usuario no encontrado",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            },
            "500": {
              "description": "Error del servidor",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        },
        "delete": {
          "tags": ["usuarios"],
          "summary": "Elimina un usuario",
          "operationId": "deleteUsuario",
          "produces": ["application/json"],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID del usuario a eliminar",
              "required": true,
              "type": "integer",
              "format": "int64"
            }
          ],
          "responses": {
            "200": {
              "description": "Usuario eliminado",
              "schema": {
                "$ref": "#/definitions/DeleteResponse"
              }
            },
            "404": {
              "description": "Usuario no encontrado",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            },
            "500": {
              "description": "Error del servidor",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },
...
      "/lugares/usuario/{id}": {
        "get": {
          "tags": ["lugares"],
          "summary": "Obtiene todos los lugares de un usuario",
          "operationId": "getLugaresByUsuario",
          "produces": ["application/json"],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID del usuario",
              "required": true,
              "type": "integer",
              "format": "int64"
            }
          ],
          "responses": {
            "200": {
              "description": "Lista de lugares del usuario",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Lugar"
                }
              }
            },
            "500": {
              "description": "Error del servidor",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },
...
      "/robots/usuario/{id}": {
        "get": {
          "tags": ["robots"],
          "summary": "Obtiene todos los robots de un usuario",
          "operationId": "getRobotsByUsuario",
          "produces": ["application/json"],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID del usuario",
              "required": true,
              "type": "integer",
              "format": "int64"
            }
          ],
          "responses": {
            "200": {
              "description": "Lista de robots del usuario",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Robot"
                }
              }
            },
            "500": {
              "description": "Error del servidor",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },
...
    },
    "definitions": {
      "Usuario": {
        "type": "object",
        "required": ["nombre", "correo", "contrasena", "id_rol"],
        "properties": {
          "id_usuario": { "type": "integer", "format": "int64", "description": "Identificador único del usuario" },
          "nombre": { "type": "string", "description": "Nombre del usuario" },
          "apellidos": { "type": "string", "description": "Apellidos del usuario" },
          "correo": { "type": "string", "format": "email", "description": "Correo electrónico del usuario" },
          "telefono": { "type": "string", "description": "Número de teléfono del usuario" },
          "huella": { "type": "string", "description": "Huella digital para autenticación biométrica" },
          "contrasena": { "type": "string", "format": "password", "description": "Contraseña del usuario" },
          "notificaciones": { "type": "boolean", "description": "Indica si el usuario recibe notificaciones" },
          "idioma": { "type": "string", "description": "Idioma preferido del usuario" },
          "volumen": { "type": "integer", "format": "int32", "minimum": 0, "maximum": 100, "description": "Volumen preferido para alertas (0-100)" },
          "activo": { "type": "boolean", "description": "Indica si la cuenta del usuario está activa" },
          "foto_perfil": { "type": "string", "description": "URL de la foto de perfil del usuario" },
          "id_rol": { "type": "integer", "format": "int64", "description": "ID del rol asignado al usuario" }
        }
      },
...
      "Lugar": {
        "type": "object",
        "required": ["id_usuario", "nombre", "direccion"],
        "properties": {
          "id_lugar": { "type": "integer", "format": "int64", "description": "Identificador único del lugar" },
          "id_usuario": { "type": "integer", "format": "int64", "description": "ID del usuario propietario del lugar" },
          "nombre": { "type": "string", "description": "Nombre del lugar" },
          "direccion": { "type": "string", "description": "Dirección del lugar" },
          "favorito": { "type": "boolean", "description": "Indica si el lugar está marcado como favorito" },
          "latitud": { "type": "number", "format": "float", "description": "Latitud geográfica del lugar" },
          "longitud": { "type": "number", "format": "float", "description": "Longitud geográfica del lugar" }
        }
      },
      "Robot": {
        "type": "object",
        "required": ["nombre", "modelo", "id_usuario"],
        "properties": {
          "id_robot": { "type": "integer", "format": "int64", "description": "Identificador único del robot" },
          "nombre": { "type": "string", "description": "Nombre del robot" },
          "modelo": { "type": "string", "description": "Modelo del robot" },
          "version_firmware": { "type": "string", "description": "Versión de firmware del robot" },
          "bateria": { "type": "integer", "format": "int32", "minimum": 0, "maximum": 100, "description": "Nivel de batería del robot (0-100)" },
          "conexion": { "type": "string", "description": "Estado de conexión del robot" },
          "estado": { "type": "string", "description": "Estado operativo del robot" },
          "averias": { "type": "string", "description": "Información sobre averías del robot" },
          "ubicacion": { "type": "string", "description": "Ubicación actual del robot" },
          "velocidad_navegacion": { "type": "string", "description": "Velocidad de navegación configurada" },
          "id_usuario": { "type": "integer", "format": "int64", "description": "ID del usuario propietario del robot" }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "description": "Mensaje de error" }
        }
      },
...
      "DeleteResponse": {
        "type": "object",
        "properties": {
          "message": { "type": "string", "description": "Mensaje de confirmación" }
        }
      }
    }
  }